+++ NEWS.md
@@ -1,3 +1,9 @@
+# dod 0.1.3
+
+* `dod.amsr()` is now local, as opposed to a call to `oce::download.amsr()`.
+  Also, it queries the new URL pattern and file type (which used to be `.gz` but
+  now is `.nc`).  See https://github.com/dankelley/oce/issues/2124.
+
 # dod 0.1.2
 
 * add `age` argument to functions that download.